Fix uint conversion

+++ consensus/state.go
@@ -237,7 +237,7 @@ func (s *StateEngine) InspireConsensus(request *message.Request) error {
 	}
 	client.saveRequest(request)
 	cMsg := message.CreateConMsg(message.MTRequest, request)
-	cMsg.From = s.NodeID
+	cMsg.From = uint(s.NodeID)
 
 	if err := s.p2pWire.BroadCast(cMsg); err != nil {
 		return err
@@ -253,7 +253,7 @@ func (s *StateEngine) InspireConsensus(request *message.Request) error {
 	//log.PrePrepare = ppMsg
 	log.clientID = request.ClientID
 	cMsg = message.CreateConMsg(message.MTPrePrepare, ppMsg)
-	cMsg.From = s.NodeID
+	cMsg.From = uint(s.NodeID)
 
 	if err := s.p2pWire.BroadCast(cMsg); err != nil {
 		return err
@@ -340,7 +340,7 @@ func (s *StateEngine) idle2PrePrepare(ppMsg *message.PrePrepare) (err error) {
 		NodeID:     s.NodeID,
 	}
 	cMsg := message.CreateConMsg(message.MTPrepare, ppMsg)
-	cMsg.From = s.NodeID
+	cMsg.From = uint(s.NodeID)
 
 	if err := s.p2pWire.BroadCast(cMsg); err != nil {
 		return err
@@ -413,7 +413,7 @@ func (s *StateEngine) prePrepare2Prepare(prepare *message.Prepare) (err error) {
 		NodeID:     s.NodeID,
 	}
 	cMsg := message.CreateConMsg(message.MTCommit, commit)
-	cMsg.From = s.NodeID
+	cMsg.From = uint(s.NodeID)
 
 	if err := s.p2pWire.BroadCast(cMsg); err != nil {
 		return err
